Amongus
Among Us is basically a game of survival, where you either have to vote off all of the imposters and complete all the tasks, and the imposter has to kill all of the crewmates or stop them from completing the designated tasks to win. If you are designated a crewmate your task is to find the imposter.